Crate eosio_core

TODO docs

Structs

AccountName
TODO docs

Action
TODO docs

ActionName
TODO docs

Asset
TODO docs

Authorization
A permission

BlockTimestamp
Block timestamp as milliseconds relative to the year 2000.

Checksum160
TODO docs

Checksum256
TODO docs

Checksum512
TODO docs

CpuWeight
CPU Weight

DeferredTransactionId
TODO docs

ExtendedAsset
TODO docs

ExtendedSymbol
TODO docs

Name
TODO docs

NetWeight
Net Weight

PermissionName
TODO docs

PublicKey
TODO docs

RamBytes
RAM in bytes

ScopeName
TODO docs

Symbol
TODO docs

SymbolCode
TODO docs

TableName
TODO docs

TimePoint
Time relative to unix epoch

TimePointSec
Time relative to unix epoch

Enums

AssetDivOpError
TODO docs

AssetOpError
TODO docs

ParseAssetError
TODO docs

ParseNameError
An error which can be returned when parsing an EOSIO name.

ParseSymbolError
An error which can be returned when parsing an EOSIO symbol.

Traits

CheckedAdd
TODO docs

CheckedDiv
TODO docs

CheckedMul
TODO docs

CheckedRem
TODO docs

CheckedSub
TODO docs

ToAction
TODO docs

Functions

bool_from_u8
TODO docs

bool_to_u8
TODO docs

f64_from_string
TODO docs

u64_from_string
TODO docs